+++ ichnaea/async/config.py
@@ -94,6 +94,10 @@ def configure_data(redis_client):
         'update_wifi': DataQueue('update_wifi', redis_client,
                                  queue_key='update_wifi'),
     }
+    for shard_id in ['%x' % i for i in range(16)]:
+        name = 'update_wifi_' + shard_id
+        data_queues[name] = DataQueue(
+            name, redis_client, queue_key=name)
     return data_queues