+++ ichnaea/data/tasks.py
@@ -143,11 +143,12 @@ def update_cell(self, batch=1000):
 
 
 @celery_app.task(base=BaseTask, bind=True, queue='celery_wifi')
-def update_wifi(self, batch=1000):
+def update_wifi(self, batch=1000, shard_id=None):
     with self.redis_pipeline() as pipe:
         with self.db_session() as session:
             station.WifiUpdater(
                 self, session, pipe,
+                shard_id=shard_id,
                 remove_task=None,
                 update_task=update_wifi,
             )(batch=batch)